[libc++][NFC] Refactor `[[nodiscard]]` tests
Originally some `[[nodiscard]]` tests were implemented in `*/test/libcxx/diagnostics`. The Standard has a library `Diagnostics` and this folder should be reserved for it by convention. Many new tetss were added to their respective subfolders. This patch moves around the already implemented `[[nodiscard]]` tests to their respective folders where they belong and standardizes the name to `nodiscard.verify.cpp` whereever possible.

N.B. This refactors only tests, which were merged. The remaining will be moved in a future patch to reduce merge conflicts on their own.
