+++ libcxx/test/libcxx/containers/container.adaptors/stack/nodiscard.verify.cpp
