+++ libcxx/test/libcxx/containers/sequences/array/nodiscard.verify.cpp
