+++ libcxx/test/libcxx/language.support/nodiscard.verify.cpp
@@ -16,6 +16,7 @@
 #include <coroutine>
 #include <exception>
 #include <initializer_list>
+#include <new>
 #include <typeinfo>
 #include <typeindex>
 
@@ -157,6 +158,46 @@ void test() {
     const std::bad_typeid bt;
 
     bc.what(); // expected-warning {{ignoring return value of function declared with 'nodiscard' attribute}}
+  }
+#endif
+
+  {
+    std::bad_alloc ex;
+
+    ex.what(); // expected-warning {{ignoring return value of function declared with 'nodiscard' attribute}}
+  }
+  {
+    std::bad_array_new_length ex;
+
+    ex.what(); // expected-warning {{ignoring return value of function declared with 'nodiscard' attribute}}
+  }
+
+  {
+    // expected-warning@+1 {{ignoring return value of function declared with 'nodiscard' attribute}}
+    ::operator new(0);
+    // expected-warning@+1 {{ignoring return value of function declared with 'nodiscard' attribute}}
+    ::operator new(0, std::nothrow);
+    // expected-warning@+1 {{ignoring return value of function declared with 'nodiscard' attribute}}
+    ::operator new[](0);
+    // expected-warning@+1 {{ignoring return value of function declared with 'nodiscard' attribute}}
+    ::operator new[](0, std::nothrow);
+#if _LIBCPP_HAS_ALIGNED_ALLOCATION
+    // expected-warning@+1 {{ignoring return value of function declared with 'nodiscard' attribute}}
+    ::operator new(0, std::align_val_t{1});
+    // expected-warning@+1 {{ignoring return value of function declared with 'nodiscard' attribute}}
+    ::operator new(0, std::align_val_t{1}, std::nothrow);
+    // expected-warning@+1 {{ignoring return value of function declared with 'nodiscard' attribute}}
+    ::operator new[](0, std::align_val_t{1});
+    // expected-warning@+1 {{ignoring return value of function declared with 'nodiscard' attribute}}
+    ::operator new[](0, std::align_val_t{1}, std::nothrow);
+#endif // _LIBCPP_HAS_ALIGNED_ALLOCATION
+  }
+
+#if TEST_STD_VER >= 17
+  {
+    int* ptr = nullptr;
+
+    std::launder(ptr); // expected-warning {{ignoring return value of function declared with 'nodiscard' attribute}}
   }
 #endif
 }
