+++ libcxx/test/libcxx/utilities/function.objects/nodiscard.verify.cpp
